replace gossiplog/opfs with gossiplog/sqlite-wasm
